backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
pytrends==4.9.2
openai==1.3.0
textblob==0.17.1
python-dotenv==1.0.0